Create Proposal for security deposit
Maybe this is not the place for this, but I would totally suggest that if a security deposit is coming directly from Bisq wallet, those funds get blocked instead of having to actually send them to the network. 
It would make for way faster processing, and it might help mitigating the current miscalculation of correct mining fee that creates unnecessary (and serious) delays on trades.
I personally cancelled one of my transactions since I got tired of having to wait and keep checking on the security deposit transaction. 
The trade was delayed for so long even before having to actually start the real trade.
Thanks
On top of that, it would encourage users to always keep some funds on their Bisq wallet.